Revert "Conditional import of ExitStack"
This reverts commit ca41d813ed0007c8ea6743fd6332447b003a2390.
